Always return the full 'qualified' URL for external routes.
Currently Pyramid 1.5a+ has a notion of "External Routes" which must
be generated by calling request.route_url() since external URL's are
fully qualified by definition. However if one is using Akhet's
URLGenerator.route() and prefers to use url paths (as pylons did by
default and is easier to deploy through proxies) there is a problem if
it is unknown if the URLs being generated are local or 'external'. For
example if one is looping through a set of routes to generate a menu
but some of the items in the menu point to external sites while most
are local.  It makes more since to me that since all external routes
need to be fully qualified (again, by definition) that the thing to do
is just return the fully qualified URL always and not an error. That
is what this patch does (hopefully.)