Reverting back to non pep-440 compliant way
The build tools are coverting the + (plus) between the public number and
the local number (as in 2.0+edgy1) to a - (dash) making it 2.0-edgy1.
Until I fifure out what I am suppsed to do here I am just reverting
back to the 'it works' version numbering scheme. I am not placing
these versions on PyPi anyways.